Removed the .babel extension
