+++ package.json
@@ -17,7 +17,6 @@
     "gulp": "gulpjs/gulp#4.0",
     "gulp-imagemin": "^3.1.1",
     "gulp-postcss": "^6.3.0",
-    "gulp-rename": "^1.2.2",
     "postcss-cssnext": "^2.9.0"
   },
   "devDependencies": {
